inference_service/requirements.txt
fastapi
uvicorn
ultralytics
pillow
requests
sqlalchemy
psycopg2-binary
python-dotenv
